preparing infra to build lib in separate
