+++ hyperon_das_atomdb_cpp/Makefile
@@ -1,11 +1,12 @@
 format:
-	@find ./src \
+	@find ./lib ./nanobind \
 		-iname "*.hpp" -o \
 		-iname "*.cpp" -o \
 		-iname "*.hh" -o \
 		-iname "*.cc" -o \
 		-iname "*.h" -o \
 		-iname "*.c" \
+		| grep -Ev '(./lib/build|./nanobind/build)' \
 		| xargs clang-format -verbose -style=file -i
 
 wheeler-images:
