+++ hyperon_das_atomdb_cpp/deps/requirements.txt
@@ -2,3 +2,5 @@ wheel>=0.37.1
 nanobind>=2.1.0
 scikit-build-core>=0.10
 auditwheel>=6.1.0
+patchelf>=0.17.2.1
+
