+++ hyperon_das_atomdb_cpp/lib/include/database.h
@@ -33,8 +33,8 @@
 #include "constants.h"
 #include "document_types.h"
 #include "exceptions.h"
+#include "expression_hasher.h"
 #include "type_aliases.h"
-#include "utils/expression_hasher.h"
 
 using namespace std;
