+++ hyperon_das_atomdb_cpp/lib/include/patterns.h
