--- hyperon_das_atomdb_cpp/src/.clang-format-ignore
@@ -1,1 +1,0 @@
-
